requirements.txt
# Core Bot Dependencies
python-telegram-bot>=20.0
pandas>=1.5.0

# Environment and Configuration
python-dotenv>=1.0.0
pytz>=2023.3

# Database and Caching  
sqlalchemy>=1.4.0,<2.0.0
redis>=4.5.0
alembic>=1.11.0

# Content Processing
nltk>=3.8
textblob>=0.17.1
profanity-check>=1.0.3
langdetect>=1.0.9

# Analytics and Monitoring
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
psutil>=5.9.0

# Security and Validation
cryptography>=41.0.0
bcrypt>=4.0.0
validators>=0.20.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.0
aiofiles>=23.0.0
pillow>=10.0.0
openpyxl>=3.1.0
schedule>=1.2.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
